renaming and designing the FloatingLabelInput
