+++ components/FloatingLabelInput.tsx
@@ -1,4 +1,5 @@
 import { useEffect, useRef, useState } from "react";
+import { COLORS } from "../theme";
 import {
   Text,
   TextInput,
@@ -7,7 +8,6 @@ import {
   Easing,
   TouchableWithoutFeedback,
 } from "react-native";
-import styles from "../stylesheets/styleFloatingInput";
 
 type Props = React.ComponentProps<typeof TextInput> & {
   label: string;
@@ -16,7 +16,7 @@ type Props = React.ComponentProps<typeof TextInput> & {
   value: string;
 };
 
-const FloatingTextInput: React.FC<Props> = (props) => {
+const FloatingLabelInput: React.FC<Props> = (props) => {
   const {
     label,
     errorText,
@@ -40,16 +40,16 @@ const FloatingTextInput: React.FC<Props> = (props) => {
     }).start();
   }, [focusAnim, isFocused, value]);
 
-  let color = isFocused ? "#080F9C" : "#A9A9A9";
+  let color = isFocused ? COLORS.primary : COLORS.lightgray;
   if (errorText) {
-    color = "#FF2D55";
+    color = COLORS.red;
   }
 
   return (
-    <View style={{ marginVertical: 10 }}>
+    <View className="my-2">
       <TextInput
+      className="border p-4 rounded-xl mx-10"
         style={[
-          styles.input,
           {
             borderColor: color,
           },
@@ -69,8 +69,8 @@ const FloatingTextInput: React.FC<Props> = (props) => {
       />
       <TouchableWithoutFeedback onPress={() => inputRef.current?.focus()}>
         <Animated.View
+          className="absolute mx-10 bg-white px-2"
           style={[
-            styles.floatingLabelContainer,
             {
               transform: [
                 {
@@ -82,7 +82,7 @@ const FloatingTextInput: React.FC<Props> = (props) => {
                 {
                   translateY: focusAnim.interpolate({
                     inputRange: [0, 1],
-                    outputRange: [16, -12],
+                    outputRange: [11, -16],
                   }),
                 },
                 {
@@ -96,10 +96,10 @@ const FloatingTextInput: React.FC<Props> = (props) => {
           ]}
         >
           <Text
+          className="text-base"
             style={[
-              styles.label,
               {
-                color,
+                color
               },
             ]}
           >
@@ -108,9 +108,9 @@ const FloatingTextInput: React.FC<Props> = (props) => {
           </Text>
         </Animated.View>
       </TouchableWithoutFeedback>
-      {!!errorText && <Text style={styles.error}>{errorText}</Text>}
+      {!!errorText && <Text className="text-sm text-red my-6">{errorText}</Text>}
     </View>
   );
 };
 
-export default FloatingTextInput;
+export default FloatingLabelInput;
